PSBJBM-33 PSBJBM-28 Subclassed the Application
Created the FreeTimeApplication class that extends Application, and use
this as the entry point for our application.

FreeTimeApplication binds to the FreeTimeCalendarService. All
theactivities in our application now get access to the
FreeTimeCalendarService through the FreeTimeApplication

Also created an ImportCalendarActivity that uses a SimpleCursorAdapter
to display a list of Calendars available on the User's device.

Importing Calendars is still not implemented