PSBJBM-9 Can now search for empty time slots
Implemented the public void findUnoccupiedTimeSlots(long beginRange,
long endRange) method in the FreeTimeCalendarService, which looks for
unoccupied time slots in the given time range, and saves the found
unoccupied slots to the unoccupiedtime table in the FreeTimeDb.db

The implementation isn't complete:

Right now it should only detect the the first unoccupied slot in the
given time range. There are three cases:

There are no events in the given time range => The whole time range is a
single unoccupied time slot.

There is at least one event in the given time range =>
if the event starts at the same time as the given time range starts,
then the first empty slot starts at the Instance.END, and ends at the
Instance.START of the NEXT Instance (if there is one. Otherwise the
empty slot ends at the end of the given time range)
otherwise, the first empty slot starts at the same time as the start of
the given time range, and the empty slot ends at the Instance.START