Change method signatures in FreeTimeCalendarService
The FreeTimeService.findUnoccupiedTimeSlots method now requires a
FreeTimeDatabaseHelper object to be passed in its parameters. This is so
that when the method is called in a Test, a new Database is created and
that database is used instead.

When this method is called from any other class, this parameter should
be set to Null.

This is subject to change (depending on if we can get the Robolectric
framework to work, and if we can find another solution for running the
tests on a separate database)