PSBJBM-55 Created java project for coding algorithm
Created the freetime-algorithm project for coding and testing the
algorithm for splitting tasks into events and spreading them out over
the remaining time before the due date.

What works: The Event model is complete (I think). the
detectEmptySlots() method works correctly (need to write a JUnit test
for it).  The GUI interface is functional. To make tests faster, events
can be created and saved in the resources/events.txt file, and can be
imported by clicking the import evens button on the GUI.

The detectEmptySlots() can be tested from the GUI (to test for a whole
day, put the start time to 00:00 and the end time to 23:59)

Currently working on the Optimiser.addTaskToCalendar() method.