PSBJBM-55 Found a bug in the optimiser
When the optimiser calls detectEmptySlots(cal, now, latestEndDate); the
EmptySlot detection doesn't work as expected. Perhaps it would be a
better idea to make repeated calls to detectEmptySlots, in 24 hour
blocks.

To do this, the following changes need to be made :
the detectEmptySlots() method should not clear the EmptySlots list each
time. Let the addTaskToCalendar() method handle that instead.

Also, once that is changed, in order to make sure that the
EmptySlotsDisplay panel on the GUI keeps showing correct results, make
the clear display button trigger the clearing of the EmptySlots list.

Also in this commit: added more comments to the classes in the model and
optimiser packages